--- node/stagehand/task/.gitignore
@@ -1,7 +1,0 @@
-.env
-node_modules
-tmp
-downloads
-.DS_Store
-dist
-cache.json
