--- node/stagehand/task/package.json
@@ -1,26 +1,0 @@
-{
-  "name": "task",
-  "type": "module",
-  "scripts": {
-    "build": "tsc",
-    "start": "tsx index.ts",
-    "postinstall": "playwright install"
-  },
-  "dependencies": {
-    "@browserbasehq/sdk": "^2.6.0",
-    "@browserbasehq/stagehand": "^2.2.1",
-    "@playwright/test": "^1.49.1",
-    "boxen": "^8.0.1",
-    "chalk": "^5.3.0",
-    "dotenv": "^16.4.7",
-    "zod": "^3.22.4",
-    "ai": "^4.3.16",
-    "@ai-sdk/openai": "^1.3.22",
-    "openai": "^4.100.0"
-  },
-  "devDependencies": {
-    "tsx": "^4.19.2",
-    "typescript": "^5.0.0"
-  },
-  "packageManager": "pnpm@9.15.0+sha512.76e2379760a4328ec4415815bcd6628dee727af3779aaa4c914e3944156c4299921a89f976381ee107d41f12cfa4b66681ca9c718f0668fa0831ed4c6d8ba56c"
-}
