--- node/stagehand/task/tsconfig.json
@@ -1,17 +1,0 @@
-{
-  "compilerOptions": {
-    "target": "ES2022",
-    "module": "NodeNext",
-    "moduleResolution": "NodeNext",
-    "outDir": "./dist",
-    "rootDir": "./",
-    "strict": true,
-    "esModuleInterop": true,
-    "skipLibCheck": true,
-    "forceConsistentCasingInFileNames": true,
-    "allowImportingTsExtensions": false,
-    "noEmit": false
-  },
-  "include": ["*.ts", "api", "examples/run.ts"],
-  "exclude": ["node_modules", "dist"]
-}
